Hash once instead of twice
`seen.insert` returns a bool depending on whether the value is present
in the HashSet. We can use this return value instead of `seen.contains`
to save one hash.

Thanks @mmun for the protip!

Also, change the return type for `main` to be more idiomatic. This will
let us preserve the error type instead of forcing all errors to be io
errors. (From:
https://www.reddit.com/r/rust/comments/8ilg97/small_tip_on_new_main_result_behavior/)